+++ src/testmod/java/ladysnake/satinrenderlayer/SatinRenderLayerTest.java
@@ -32,7 +32,6 @@ import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.entity.EntityType;
@@ -43,36 +42,37 @@ import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import net.minecraft.world.World;
+import org.jetbrains.annotations.NotNull;
 
 public final class SatinRenderLayerTest {
 
     /* * * * ManagedShaderEffect-based RenderLayer entity rendering * * * */
 
-    public static final EntityType<IronGolemEntity> ILLUSION_GOLEM =
+    public static final @NotNull EntityType<IronGolemEntity> ILLUSION_GOLEM =
             Registry.register(
                     Registries.ENTITY_TYPE,
-                    new Identifier("satinrenderlayer", "illusion_golem"),
-                    FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, IronGolemEntity::new).dimensions(EntityType.IRON_GOLEM.getDimensions()).build()
+                    Identifier.of("satinrenderlayer", "illusion_golem"),
+                    EntityType.Builder.create(IronGolemEntity::new, SpawnGroup.CREATURE).dimensions(EntityType.IRON_GOLEM.getWidth(), EntityType.IRON_GOLEM.getHeight()).build()
             );
 
-    public static final ManagedShaderEffect illusionEffect = ShaderEffectManager.getInstance().manage(new Identifier("satinrenderlayer", "shaders/post/illusion.json"),
+    public static final ManagedShaderEffect illusionEffect = ShaderEffectManager.getInstance().manage(Identifier.of("satinrenderlayer", "shaders/post/illusion.json"),
             effect -> effect.setUniformValue("ColorModulate", 1.2f, 0.7f, 0.2f, 1.0f));
     public static final ManagedFramebuffer illusionBuffer = illusionEffect.getTarget("final");
 
     /* * * * ManagedShaderProgram-based RenderLayer entity rendering * * * */
 
-    public static final EntityType<WitherEntity> RAINBOW_WITHER =
+    public static final @NotNull EntityType<WitherEntity> RAINBOW_WITHER =
             Registry.register(
                     Registries.ENTITY_TYPE,
-                    new Identifier("satinrenderlayer", "rainbow_wither"),
-                    FabricEntityTypeBuilder.create(SpawnGroup.CREATURE, (EntityType<WitherEntity> entityType, World world) -> {
+                    Identifier.of("satinrenderlayer", "rainbow_wither"),
+                    EntityType.Builder.create((EntityType<WitherEntity> entityType, World world) -> {
                         WitherEntity witherEntity = new WitherEntity(entityType, world);
                         witherEntity.setAiDisabled(true);
                         return witherEntity;
-                    }).dimensions(EntityType.WITHER.getDimensions()).build()
+                    }, SpawnGroup.CREATURE).dimensions(EntityType.WITHER.getWidth(), EntityType.WITHER.getHeight()).build()
             );
 
-    public static final ManagedCoreShader rainbow = ShaderEffectManager.getInstance().manageCoreShader(new Identifier("satinrenderlayer", "rainbow"));
+    public static final ManagedCoreShader rainbow = ShaderEffectManager.getInstance().manageCoreShader(Identifier.of("satinrenderlayer", "rainbow"));
     private static final Uniform1f uniformSTime = rainbow.findUniform1f("STime");
 
     private static int ticks;